moved select/index to query/index
